Move functionality from connector into main function with hooks
