+++ frontend/src/Indexer/Add/AddIndexerModalContent.js
@@ -1,5 +1,7 @@
 import PropTypes from 'prop-types';
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { createSelector } from 'reselect';
 import Alert from 'Components/Alert';
 import TextInput from 'Components/Form/TextInput';
 import Button from 'Components/Link/Button';
@@ -14,6 +16,8 @@ import TableBody from 'Components/Table/TableBody';
 import { kinds, scrollDirections } from 'Helpers/Props';
 import getErrorMessage from 'Utilities/Object/getErrorMessage';
 import translate from 'Utilities/String/translate';
+import { fetchIndexerSchema, SELECT_INDEXER_SCHEMA, SET_INDEXER_SCHEMA_SORT } from '../../Store/Actions/indexerActions';
+import createClientSideCollectionSelector from '../../Store/Selectors/createClientSideCollectionSelector';
 import SelectIndexerRow from './SelectIndexerRow';
 import styles from './AddIndexerModalContent.css';
 
@@ -44,23 +48,57 @@ const columns = [
   }
 ];
 
+function createMapStateToProps() {
+  return createSelector(
+    createClientSideCollectionSelector('indexers.schema'),
+    (indexers) => {
+      const {
+        isFetching,
+        isPopulated,
+        error,
+        items,
+        sortDirection,
+        sortKey
+      } = indexers;
+
+      return {
+        isFetching,
+        isPopulated,
+        error,
+        indexers: items,
+        sortKey,
+        sortDirection
+      };
+    }
+  );
+}
+
 function AddIndexerModalContent({
-  indexers,
-  onIndexerSelect,
-  sortKey,
-  sortDirection,
-  isFetching,
-  isPopulated,
-  error,
-  onSortPress,
   onModalClose
 }) {
   //
   // Lifecycle
   const [filter, setFilter] = useState('');
+  const dispatch = useDispatch();
+
+  //
+  // Data
+  const { indexers, isFetching, isPopulated, error, sortKey, sortDirection } = useSelector(createMapStateToProps());
 
   //
   // Listeners
+  useEffect(() => {
+    dispatch(fetchIndexerSchema());
+  }, []);
+
+  function onSortPress(eventSortKey) {
+    dispatch({ type: SET_INDEXER_SCHEMA_SORT, payload: { sortKey: eventSortKey } });
+  }
+
+  function onIndexerSelect({ implementation, name }) {
+    dispatch({ type: SELECT_INDEXER_SCHEMA, payload: { implementation, name } });
+    onModalClose({ indexerSelected: true });
+  }
 
   function onFilterChange({ value }) {
     setFilter(value);
@@ -88,7 +126,7 @@ function AddIndexerModalContent({
           placeholder={translate('FilterPlaceHolder')}
           name="filter"
           value={filter}
-          autoFocus
+          autoFocus={true}
           onChange={onFilterChange}
         />
 
@@ -153,14 +191,6 @@ function AddIndexerModalContent({
 }
 
 AddIndexerModalContent.propTypes = {
-  isFetching: PropTypes.bool.isRequired,
-  isPopulated: PropTypes.bool.isRequired,
-  error: PropTypes.object,
-  sortKey: PropTypes.string,
-  sortDirection: PropTypes.string,
-  onSortPress: PropTypes.func.isRequired,
-  indexers: PropTypes.arrayOf(PropTypes.object).isRequired,
-  onIndexerSelect: PropTypes.func.isRequired,
   onModalClose: PropTypes.func.isRequired
 };
